Refactored the game content into a separate file for eaiser use
