+++ src/pages/Recover/PasswordInput.css
@@ -2,14 +2,6 @@
   box-sizing: border-box;
 }
 
-body {
-  height: 100%;
-  margin: 0;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
 .fullscreen-container {
   display: flex;
   justify-content: center;
