+++ src/pages/Recover/Recover.css
@@ -1,9 +1,3 @@
-body {
-  margin: 0;
-  font-family: Arial, sans-serif;
-  background-color: #abb4c3;
-}
-
 * {
   box-sizing: border-box;
 }
